Added tests for filtering
